partner config options
Created PartnerConfig class that handles user selection of a partner.
As of now, the info gets stored in the SESSION (see
includes/set_api_info.php and flashfoto/config.php) but maybe we can
think of a better way.

Still need to test this on one of the ads access servers.